cloud_functions/sync_clients_to_gcs/requirements.txt
google-cloud-firestore==2.13.0
google-cloud-storage==2.10.0
google-cloud-functions-framework==3.5.0
firebase-admin==6.2.0

